Keep three 256K Flash-Next sessions within the memory budget
The validated three-slot profile cannot fit main-attention QSA K/V in BF16 with a safe margin. Store those rows in row-scaled FP8 E4M3 by default, keep BF16 and INT8 as explicit A/B options, and use a dtype-aware indexed write because CUDA index_copy_ lacks Float8 support. Eager MTP teacher-forced sync grows temporary rows only when graph capture is not active, preserving fixed graph addresses. Update the runbook and quality probe for the reproducible 3x256K profile.

Constraint: One SM120 GPU with approximately 96 GiB and three independent 256K sessions
Rejected: Lower model or INT4 weight precision | the request forbids changing the weight format and the model weights remain NVFP4
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep BF16/INT8 explicit via QSR_FLASHNEXT_QSA_KV_DTYPE and repeat a cold OOM and quality gate before changing slot or context counts
Tested: ruff check; 28 QSA tests; 56 Flash-Next CPU tests with 10 CUDA skips; FP8 CUDA Graph and MTP verify; three concurrent HTTP smoke requests on 8300
Not-tested: Full long-context semantic benchmark under sustained three-way load

+++ scripts/fn6_spec_batch.py
@@ -1377,7 +1377,9 @@ def main() -> None:
         },
     }
     for name, state in speculative_states.items():
-        bad = int((~torch.isfinite(state)).sum())
+        # CUDA does not provide ``isfinite`` for Float8 E4M3.  Check the
+        # dequantized view so the FP8 cache quality gate remains active.
+        bad = int((~torch.isfinite(state.float())).sum())
         if bad:
             raise RuntimeError(f"speculative state {name} produced {bad} non-finite values")
 
@@ -1411,8 +1413,8 @@ def main() -> None:
     state_max_abs = []
 
     def record_state(name: str, got: torch.Tensor, expected: torch.Tensor) -> None:
-        got_bad = int((~torch.isfinite(got)).sum())
-        expected_bad = int((~torch.isfinite(expected)).sum())
+        got_bad = int((~torch.isfinite(got.float())).sum())
+        expected_bad = int((~torch.isfinite(expected.float())).sum())
         if got_bad or expected_bad:
             raise RuntimeError(
                 f"state comparison {name} encountered non-finite values: "
